Ensure drop amounts default to 1 and prevent empty rows

+++ SFServer.API/Controllers/PurchasesController.cs
@@ -233,13 +233,13 @@ public class PurchasesController : ControllerBase
                         Id = Guid.CreateVersion7(),
                         UserId = userId,
                         ItemId = drop.TargetId,
-                        Amount = (int)drop.Amount
+                        Amount = drop.Amount
                     };
                     _db.PlayerInventoryItems.Add(inv);
                 }
                 else
                 {
-                    inv.Amount += (int)drop.Amount;
+                    inv.Amount += drop.Amount;
                 }
             }
             else if (drop.Type == ProductDropType.Currency)
